Do not load fixes.zcml automatically on Plone 3.
This relied on z3c.jbot, which may not be a good idea on Plone 3.
Instead, in the tests we install our GS profile when we are on Plone 3.

This does mean that on Plone 3 we skip the tests that check that a
completely missing honeypot field gives an error: we cannot be certain
that our fixes have been loaded, so we do not update the list of extra
protected actions on startup.