Split honeypot_field and authenticator view.
When `@@authenticator`` is called, return the view without rendering it.
We had defined a template in zcml, so unlike the original we used to
return a rendered template when we were called with
``tal:define="authenticator context/@@authenticator"``.

Fixes AttributeError: 'unicode' object has no attribute 'token', for
example on PloneFormGen quickedit form.