Started on Plone 5.
The contact-info tests pass now, after lots of struggling.

Several other tests need to be adapted to Plone 5, which uses other
form and input names, and needs an argument when calling form.submit,
otherwise the page basically just gets reloaded.

Several tests just need to be ignored on Plone 5, because the skin
template or script that we are testing is gone.

Our @@authenticator override works, but it does not always look like
that.  The plone.protect auto protection transform adds the original
authenticator *without* our honeypot to forms that do not yet have an
authenticator.  I added a transform override in auto.py to add our
honeypot as well, but I commented out the zcml.  Problem is that the
transform is only active for authenticated users, and we also need if
for anonymous users, so this does not really help.

So in Plone 5, we need to add our honeypot field to the titlelessform
macro in plone/app/z3cform/templates/macros.pt
That solves it for lots of forms in one go.

Current test status: 83 tests, 3 failures, 29 errors.  That is down
from about 70 failures or errors, so we have progress. :-)